✨ (#88) 좋아요 화면 렌더링 조건 처리
- 좋아요 데이터가 있을 시 리스트 컴포턴트 보이게
